Rename 1P rules file
