+++ smdebug_rulesconfig/__init__.py
@@ -1,22 +1,21 @@
-from ._1PRules import vanishing_gradient
-from ._1PRules import all_zero
-from ._1PRules import check_input_images
-from ._1PRules import similar_across_runs
-from ._1PRules import weight_update_ratio
-from ._1PRules import exploding_tensor
-from ._1PRules import unchanged_tensor
-from ._1PRules import loss_not_decreasing
-from ._1PRules import check_input_images
-from ._1PRules import dead_relu
-from ._1PRules import confusion
-from ._1PRules import class_imbalance
-from ._1PRules import confusion
-from ._1PRules import overfit
-from ._1PRules import tree_depth
-from ._1PRules import tensor_variance
-from ._1PRules import overtraining
-from ._1PRules import poor_weight_initialization
-from ._1PRules import saturated_activation
-from ._1PRules import nlp_sequence_ratio
+from .builtin_rules import vanishing_gradient
+from .builtin_rules import all_zero
+from .builtin_rules import check_input_images
+from .builtin_rules import similar_across_runs
+from .builtin_rules import weight_update_ratio
+from .builtin_rules import exploding_tensor
+from .builtin_rules import unchanged_tensor
+from .builtin_rules import loss_not_decreasing
+from .builtin_rules import dead_relu
+from .builtin_rules import confusion
+from .builtin_rules import class_imbalance
+from .builtin_rules import confusion
+from .builtin_rules import overfit
+from .builtin_rules import tree_depth
+from .builtin_rules import tensor_variance
+from .builtin_rules import overtraining
+from .builtin_rules import poor_weight_initialization
+from .builtin_rules import saturated_activation
+from .builtin_rules import nlp_sequence_ratio
 
 from ._collections import get_collection
